Move build utils to test utils
